requirements.txt
asgiref==3.7.2
boto3==1.40.0
botocore==1.40.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==3.14.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==2.6.0
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==11.0.0
Faker==24.4.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
marshmallow==3.21.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.10
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.13.1
setuptools==69.2.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==8.10.0
typing_extensions==4.11.0
tzdata==2023.4
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
